texture_cache: Reintroduce preserve_contents accurately
This reverts commit 94b0e2e5dae4e0bd0021ac2d8fe1ff904a93ee69.

preserve_contents proved to be a meaningful optimization. This commit
reintroduces it but properly implemented on OpenGL.

We have to make sure the clear removes all the previous contents of the
image.

It's not currently implemented on Vulkan because we can do smart things
there that's preferred to be introduced in a separate commit.