dist: add AppStream metadata
The AppStream file is mostly copied from the one already used by the
Flatpak yuzu build:
https://github.com/flathub/org.yuzu_emu.yuzu/blob/62fc225acfa0178ddf61be68ad9be60841077263/org.yuzu_emu.yuzu.metainfo.xml

As it already defines the application id as org.yuzu_emu.yuzu I renamed
the yuzu.desktop and yuzu.xml files so that they match.

I've also made some minor tweaks to it, like fixing the capitalization
of "yuzu", adding a few keys and sorting them as presented in the
documentation.

Lastly, I added PrefersNonDefaultGPU=true to the .desktop file so that
yuzu is launched with the dedicated graphics card on Linux.